plateau モジュールの外部パッケージ化
