+++ README.md
@@ -1,21 +1,20 @@
-# plateau-qgis-plugin-rev2
+# plateau-qgis-plugin
 
-[PoC]
+A QGIS plugin for loading PLATEAU 3D city models. (PLATEAU 3D 都市モデルを読み込むための QGIS プラグイン)
+
+## 開発
 
 - QGIS 3.28 で動作させるため Python 3.9 の文法で実装する必要があります。
-- `plateau_plugin.plateau` モジュールは、将来切り出して QGIS 以外でも使うことを意図しています。QGIS に依存しない汎用的なパッケージとして実装してください。
+- PLATEAU データのパーサなどは [plateau-py](https://github.com/MIERUNE/plateau-py) パッケージとして分離されています。
 
-## 実行
-
-QGIS を使わずにテスト実行:
-
-```console
-cd plateau_plugin
-python3 -m plateau /path/to/city/udx/luse/563846_luse_6668_op.gml
-```
-
-QGIS にデプロイ (macOS):
+QGIS にデプロイする:
 
 ```console
 make deploy
+```
+
+最新の `plateau-py` を取り込む:
+
+```console
+make update_dependencies
 ```
