+++ plateau_plugin/plateau/parse/__init__.py
@@ -1,0 +1,3 @@
+from .parser import CityObjectParser, ParserSettings, PlateauCityGmlParser
+
+__all__ = ["ParserSettings", "PlateauCityGmlParser", "CityObjectParser"]
